+++ crypto/Cargo.toml
@@ -1,75 +1,19 @@
 [package]
-name = "xdag-plus"
+name = "crypto"
 version = "0.1.5"
 edition = "2021"
 authors = ["swordlet <swordlet@qq.com>"]
-build = "build.rs"
-description = "a xdag cross-platform gui wallet powered by rust lang and slint-ui"
+description = "xdag cross-platform gui wallet crypto"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-path = "./src/main.rs"
-name = "xdag-plus"
 
 [lib]
 path = "./src/lib.rs"
 crate-type = ["lib", "cdylib"]
-name = "xdagplus_lib"
+name = "crypto"
 
 [dependencies]
-slint = { version = "1.10.0", features = [
-    "backend-winit",
-    "renderer-femtovg",
-    "compat-1-2",
-], default-features = false }
-rayon = "1.10.0"
-i-slint-backend-winit = { version = "1.10.0", features = ["renderer-femtovg"] }
-winit = "0.30.9"
-
-bip32 = { path = "../bip32" }
-rand_core = { version = "0.6.4", features = ["getrandom"] }
 ripemd = "0.1.3"
 sha2 = "0.10.8"
-bs58 = { version = "0.5.1", features = ["check"] }
-rand = "0.8.5"
-bcrypt = "0.16.0"
 aes = "0.8.4"
 cbc = "0.1.2"
-
-tracing = "0.1.41"
-tracing-subscriber = { version = "0.3.19", features = [
-    "json",
-    "env-filter",
-    "fmt",
-    "time",
-    "local-time",
-] }
-time = "0.3.37"
-tracing-appender = "0.2.3"
-
-anyhow = "1.0.95"
-thiserror = "2.0.9"
-
-byteorder = "1.5.0"
-hex = "0.4.3"
-
-reqwest = { version = "0.12.15", features = [
-    "json",
-    "rustls-tls-webpki-roots",
-], default-features = false }
-tokio = { version = "1.38.0", features = ["macros", "rt-multi-thread"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-jsonrpsee = { version = "0.24.0", features = ["http-client"] }
-
-directories = "6.0"
-toml = "0.8.20"
-qrcode = "0.14.1"
-open = "=5.3.2"
-rustls = "0.23.26"
-rustls-webpki = "0.102"
-
-[build-dependencies]
-slint-build = "1.10.0"
-winres = "0.1.12"
