--- crypto/src/mod.rs
@@ -1,6 +1,0 @@
-mod aes;
-mod hash;
-pub use self::aes::aes_cbc_decrypt;
-pub use self::aes::aes_cbc_encrypt;
-pub use self::hash::hash160;
-pub use self::hash::sha256;
