+++ xerror/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "xerror"
+version = "0.1.5"
+edition = "2021"
+authors = ["swordlet <swordlet@qq.com>"]
+description = "xdag cross-platform gui wallet errors"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+path = "./src/lib.rs"
+crate-type = ["lib", "cdylib"]
+name = "xerror"
+
+[dependencies]
+bip32 = { path = "../bip32" }
+bs58 = { version = "0.5.1", features = ["check"] }
+thiserror = "2.0.9"
+reqwest = { version = "0.12.15", features = [
+    "json",
+    "rustls-tls-webpki-roots",
+], default-features = false }
+# jsonrpsee = { version = "0.24.0", features = ["http-client"] }
+toml = "0.8.20"
